Adding inline Util::size2int helper

+++ src/shop.h
@@ -75,5 +75,5 @@ private:
 	bool shop_data_outdated_;
 	bool auto_update_;
 	bool submitting_;
-	size_t requests_completed_;
+	int requests_completed_;
 };
